Added features to planning controller
Can now show the download link to the file.
Can display the planning form but requires to have the token to send
request to the server.
Therefore, need to access to this page only if the user is logged in.

Also need to manage errors on different pages and cases.
=> create an error controller ? or page ?